spark-job/requirements.txt
kafka-python==2.0.2
PyMySQL==1.1.1
